refactor: server schema
