+++ back-end-graphql/package.json
@@ -2,7 +2,7 @@
   "name": "back-end-graphql",
   "version": "1.0.0",
   "description": "gestion-app with yarn graphql react",
-  "main": "index.js",
+  "main": "server.js",
   "license": "MIT",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
