+++ back-end-graphql/src/errorHandle/error.js
@@ -1,0 +1,5 @@
+exports.errorMessage = (msg) => {
+  return {
+    message: msg,
+  };
+};
