+++ back-end-graphql/src/gestion-app.graphql
@@ -1,2 +1,1 @@
-
 # import User from "user/user.graphql"
