--- back-end-graphql/src/graphql/errorHandle/errorTypes.js
@@ -1,5 +1,0 @@
-exports.errorMessage = (msg)=>{
-    return {
-        message : msg
-    }
-}
