+++ back-end-graphql/src/graphql/typeDefs/typeDefs.js
@@ -6,7 +6,7 @@ const {
 const User = require("../../user/user.schema");
 const Team = require("../../model/Team");
 const Activity = require("../../model/Activity");
-const { errorMessage } = require("../errorHandle/errorTypes");
+const { errorMessage } = require("../../errorHandle/error");
 
 //console.log(createTypeDefsFromObject('errorResponse',errorMessage('String')))
 console.log(
