+++ back-end-graphql/src/user/user.js
@@ -1,7 +1,7 @@
 const { ApolloError } = require("apollo-server");
 const User = require("./user.schema");
 const bcryptjs = require("bcryptjs");
-const { errorMessage } = require("../graphql/errorHandle/errorTypes");
+//const { errorMessage } = require("../graphql/errorHandle/errorTypes");
 const jwt = require("jsonwebtoken");
 const { applySpec, mergeLeft } = require("ramda");
